- fixed hang in connection due not NULL paramet passed as local socket - fixed memory leak in destruction of connection - generating config file - force to chose throwing exception during build - fixed compilatin by VS2010
error C2065: '__PRETTY_FUNCTION__' : undeclared identifier	src\statement.cpp	33
error C2872: 'time' : ambiguous symbol	src\statement.cpp	163
error C4430: missing type specifier - int assumed. Note: C++ does not support default-int	D:\Workspace\git.mailserver.8.5.X\connect\extLibs\mariadb++\src\statement.cpp	163	1
error C2143: syntax error : missing ',' before '&'	src\statement.cpp	163
error C2511: 'void mariadb::statement::set_time(mariadb::u32,const int)' : overloaded member function not found in 'mariadb::statement'	src\statement.cpp	164
error C3861: 'round': identifier not found	src\decimal.cpp	25
error C2668: 'pow' : ambiguous call to overloaded function	src\decimal.cpp	32
error C2466: cannot allocate an array of constant size 0	src\decimal.cpp	32
error C2065: '__func__' : undeclared identifier	src\date_time.cpp	144


src\bind.cpp(49): warning C4800: 'const my_bool' : forcing value to bool 'true' or 'false' (performance warning)
src\connection.cpp(238): warning C4244: '+=' : conversion from 'my_ulonglong' to 'mariadb::u32', possible loss of data
src\connection.cpp(269): warning C4244: 'return' : conversion from 'my_ulonglong' to 'mariadb::u32', possible loss of data
src\date_time.cpp(65): warning C4996: 'localtime': This function or variable may be unsafe. Consider using localtime_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
src\date_time.cpp(633): warning C4996: 'localtime': This function or variable may be unsafe. Consider using localtime_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
src\date_time.cpp(647): warning C4996: 'gmtime': This function or variable may be unsafe. Consider using gmtime_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
src\date_time.cpp(673): warning C4996: 'sprintf': This function or variable may be unsafe. Consider using sprintf_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
src\date_time.cpp(675): warning C4996: 'sprintf': This function or variable may be unsafe. Consider using sprintf_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
src\date_time.cpp(684): warning C4996: 'sprintf': This function or variable may be unsafe. Consider using sprintf_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
src\date_time.cpp(493): warning C4244: 'initializing' : conversion from 'mariadb::s64' to 'mariadb::u32', possible loss of data
src\date_time.cpp(496): warning C4244: 'initializing' : conversion from 'mariadb::s64' to 'mariadb::u32', possible loss of data
src\date_time.cpp(499): warning C4244: 'initializing' : conversion from 'mariadb::s64' to 'mariadb::u32', possible loss of data
src\date_time.cpp(502): warning C4244: 'initializing' : conversion from 'mariadb::s64' to 'mariadb::u32', possible loss of data
src\date_time.cpp(586): warning C4244: 'argument' : conversion from 'mariadb::u16' to 'mariadb::u8', possible loss of data
src\date_time.cpp(658): warning C4244: 'initializing' : conversion from 'mariadb::u16' to 'mariadb::u8', possible loss of data
src\date_time.cpp(659): warning C4244: 'initializing' : conversion from 'mariadb::u16' to 'mariadb::u8', possible loss of data
src\decimal.cpp(173): warning C4244: 'argument' : conversion from 'mariadb::s64' to 'mariadb::u8', possible loss of data
src\result_set.cpp(264): warning C4800: 'mariadb::u64' : forcing value to bool 'true' or 'false' (performance warning)
src\result_set.cpp(272): warning C4244: 'return' : conversion from 'mariadb::u64' to 'mariadb::u8', possible loss of data
src\result_set.cpp(280): warning C4244: 'return' : conversion from 'mariadb::s64' to 'mariadb::s8', possible loss of data
src\result_set.cpp(288): warning C4244: 'return' : conversion from 'mariadb::u64' to 'mariadb::u16', possible loss of data
src\result_set.cpp(296): warning C4244: 'return' : conversion from 'mariadb::s64' to 'mariadb::s16', possible loss of data
src\result_set.cpp(304): warning C4244: 'return' : conversion from 'mariadb::u64' to 'mariadb::u32', possible loss of data
src\result_set.cpp(312): warning C4244: 'return' : conversion from 'mariadb::s64' to 'mariadb::s32', possible loss of data
src\save_point.cpp(33): warning C4996: 'sprintf': This function or variable may be unsafe. Consider using sprintf_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
src\save_point.cpp(35): warning C4996: 'sprintf': This function or variable may be unsafe. Consider using sprintf_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
src\save_point.cpp(49): warning C4996: 'sprintf': This function or variable may be unsafe. Consider using sprintf_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
src\save_point.cpp(64): warning C4996: 'sprintf': This function or variable may be unsafe. Consider using sprintf_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
src\statement.cpp(85): warning C4244: 'return' : conversion from 'my_ulonglong' to 'mariadb::s32', possible loss of data
src\statement.cpp(96): warning C4244: 'return' : conversion from 'my_ulonglong' to 'mariadb::u32', possible loss of data
src\statement.cpp(125): warning C4244: 'argument' : conversion from 'mariadb::u64' to 'unsigned long', possible loss of data
src\time.cpp(55): warning C4996: 'localtime': This function or variable may be unsafe. Consider using localtime_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
src\time.cpp(366): warning C4244: 'initializing' : conversion from 'mariadb::s64' to 'mariadb::u32', possible loss of data
src\time.cpp(369): warning C4244: 'initializing' : conversion from 'mariadb::s64' to 'mariadb::u32', possible loss of data
src\time.cpp(372): warning C4244: 'initializing' : conversion from 'mariadb::s64' to 'mariadb::u32', possible loss of data
src\time.cpp(375): warning C4244: 'argument' : conversion from 'mariadb::s64' to 'mariadb::u32', possible loss of data
src\time.cpp(419): warning C4996: 'localtime': This function or variable may be unsafe. Consider using localtime_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
src\time.cpp(427): warning C4996: 'gmtime': This function or variable may be unsafe. Consider using gmtime_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
src\time.cpp(438): warning C4244: 'initializing' : conversion from 'mariadb::u16' to 'mariadb::u8', possible loss of data
src\time.cpp(444): warning C4244: '=' : conversion from 'mariadb::u16' to 'mariadb::u8', possible loss of data
src\time.cpp(450): warning C4244: 'initializing' : conversion from 'double' to 'mariadb::u16', possible loss of data
src\time.cpp(460): warning C4996: 'sprintf': This function or variable may be unsafe. Consider using sprintf_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
src\time.cpp(462): warning C4996: 'sprintf': This function or variable may be unsafe. Consider using sprintf_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.

+++ include/mariadb++/concurrency.hpp
@@ -55,8 +55,8 @@ namespace mariadb
 		//
 		// Query executed, result ready to be used
 		//
-		extern s32 execute(handle handle);
-		extern u32 insert(handle handle);
+		extern u64 execute(handle handle);
+		extern u64 insert(handle handle);
 		extern result_set_ref query(handle handle);
 
 		//
